[v4] Refactor a bunch of debug code to DEBUG (#388)
