--- addon/-debug/index.js
@@ -1,4 +1,0 @@
-import DebugMixin from './edge-visualization/debug-mixin';
-import Collection from '../components/vertical-collection/component';
-
-Collection.reopen(DebugMixin);
