--- addon/-debug/utils/validate-rect.js
@@ -1,13 +1,0 @@
-import { isNonZero } from './validate-style';
-
-export function hasDimension(rect, prop) {
-  return isNonZero(rect[prop]);
-}
-
-export function hasDimensionAbove(rect, prop, amount) {
-  return hasDimension(rect, prop) && rect[prop] >= amount;
-}
-
-export function hasDimensionEqual(rect, prop, amount) {
-  return hasDimension(rect, prop) && rect[prop] === amount;
-}
