--- app/initializers/debug.js
@@ -1,6 +1,0 @@
-import '@html-next/vertical-collection/-debug';
-
-export default {
-  name: 'vertical-collection-debug',
-  initialize() {}
-};
